Moved wepapp2 fiddling to web package
