+++ api/handlers/containerhandler.py
@@ -2,7 +2,7 @@ import bson
 import datetime
 import dateutil
 
-from .. import base
+from ..web import base
 from .. import util
 from .. import config
 from .. import validators
