+++ api/handlers/devicehandler.py
@@ -1,6 +1,6 @@
 import datetime as dt
 
-from .. import base
+from ..web import base
 from .. import config
 from .. import util
 from ..auth import require_drone, require_login, require_superuser
