+++ api/handlers/grouphandler.py
@@ -1,6 +1,6 @@
 import datetime
 
-from .. import base
+from ..web import base
 from .. import util
 from .. import validators
 from ..auth import groupauth
