+++ api/handlers/reporthandler.py
@@ -2,7 +2,7 @@ import bson
 import dateutil
 import copy
 
-from .. import base
+from ..web import base
 from .. import config
