+++ api/handlers/userhandler.py
@@ -3,7 +3,7 @@ import datetime
 import pymongo
 import os
 
-from .. import base
+from ..web import base
 from .. import util
 from .. import config
 from .. import validators
